+++ eslint.config.mjs
@@ -3,7 +3,7 @@ import eslintConfigPrettier from 'eslint-config-prettier';
 import tseslint from 'typescript-eslint';
 
 export default tseslint.config(
-  { ignores: ['**/node_modules', '**/dist'] },
+  { ignores: ['**/node_modules', '**/dist', 'src/runtime/**'] },
   eslint.configs.recommended,
   tseslint.configs.recommended,
   eslintConfigPrettier